% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{arrayr}
\alias{arrayr}
\alias{arrayr-package}
\title{Manage Array Structures}
\description{
The package provides classes and methods for managing array structures.
}
\section{Core classes}{


 \itemize{
   \item{\code{\link[arrayr]{Array}}: }{

     Class for representing list-based array structures.
   }
   \item{\code{\link[arrayr]{ArrayEnvironment}}: }{

     Class for representing environment-based array structures.
   }
}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/arrayr}
}

